+++ arma3srv-script.bash
@@ -2,16 +2,21 @@
 
 #Arma 3 server script by 7thCore
 #If you do not know what any of these settings are you are better off leaving them alone. One thing might brake the other if you fiddle around with it.
-export VERSION="202001211631"
+export VERSION="202001301608"
 
 #Basics
 export NAME="Arma3Srv" #Name of the tmux session
 if [ "$EUID" -ne "0" ]; then #Check if script executed as root and asign the username for the installation process, otherwise use the executing user
 	USER="$(whoami)"
 else
-	echo "WARNING: Installation mode"
-	read -p "Please enter username (leave empty for arma):" USER #Enter desired username that will be used when creating the new user
-	USER=${USER:=arma} #If no username was given, use default
+	if [ "$1" == "-install" ] || [ "$1" == "-install_packages" ]; then
+		echo "WARNING: Installation mode"
+		read -p "Please enter username (leave empty for arma):" USER #Enter desired username that will be used when creating the new user
+		USER=${USER:=arma} #If no username was given, use default
+	else
+		echo "Error: This script, once installed, is meant to be used by the user it created and should not under any circumstances be used with sudo or by the root user for the $1 function. Only -install and -install_packages work with sudo/root. Log in to your created user (default: arma) with sudo -i -u arma and execute your script without root from the coresponding scripts folder."
+		exit 1
+	fi
 fi
 
 #Server configuration
@@ -483,6 +488,21 @@ script_install_mods() {
 
 #Check for updates. If there are updates available, shut down the server, update it and restart it.
 script_update() {
+	if [[ "$STEAMCMDUID" == "not_stored" ]] && [[ "$STEAMCMDPSW" == "not_stored" ]]; then
+		while [[ "$STEAMCMDSUCCESS" != "0" ]]; do
+			read -p "Enter your Steam username: " STEAMCMDUID
+			echo ""
+			read -p "Enter your Steam password: " STEAMCMDPSW
+			su - $USER -c "steamcmd +login $STEAMCMDUID $STEAMCMDPSW +quit"
+			STEAMCMDSUCCESS=$?
+			if [[ "$STEAMCMDSUCCESS" == "0" ]]; then
+				echo "Steam login for $STEAMCMDUID: SUCCEDED!"
+			elif [[ "$STEAMCMDSUCCESS" != "0" ]]; then
+				echo "Steam login for $STEAMCMDUID: FAILED!"
+				echo "Please try again."
+			fi
+		done
+	fi
 	echo "$(date +"%Y-%m-%d %H:%M:%S") [$VERSION] [$NAME] [INFO] (Update) Initializing update check." | tee -a "$LOG_SCRIPT"
 	if [[ "$BETA_BRANCH_ENABLED" == "1" ]]; then
 		echo "$(date +"%Y-%m-%d %H:%M:%S") [$VERSION] [$NAME] [INFO] (Update) Beta branch enabled. Branch name: $BETA_BRANCH_NAME" | tee -a "$LOG_SCRIPT"
@@ -570,6 +590,21 @@ script_update() {
 
 #Check for updates of mods. If there are updates available, shut down the server, update it and restart it.
 script_update_mods() {
+	if [[ "$STEAMCMDUID" == "not_stored" ]] && [[ "$STEAMCMDPSW" == "not_stored" ]]; then
+		while [[ "$STEAMCMDSUCCESS" != "0" ]]; do
+			read -p "Enter your Steam username: " STEAMCMDUID
+			echo ""
+			read -p "Enter your Steam password: " STEAMCMDPSW
+			su - $USER -c "steamcmd +login $STEAMCMDUID $STEAMCMDPSW +quit"
+			STEAMCMDSUCCESS=$?
+			if [[ "$STEAMCMDSUCCESS" == "0" ]]; then
+				echo "Steam login for $STEAMCMDUID: SUCCEDED!"
+			elif [[ "$STEAMCMDSUCCESS" != "0" ]]; then
+				echo "Steam login for $STEAMCMDUID: FAILED!"
+				echo "Please try again."
+			fi
+		done
+	fi
 	if [[ "$MODS_ENABLED" == "1" ]]; then
 		MODS_TO_UPDATE=""
 		echo "$(date +"%Y-%m-%d %H:%M:%S") [$VERSION] [$NAME] [INFO] (Mod Update) Checking for mod updates." | tee -a "$LOG_SCRIPT"
@@ -985,8 +1020,10 @@ script_timer_one() {
 		script_enabled
 		script_logs
 		script_autobackup
-		script_update
-		script_update_mods
+		if [[ "$STEAMCMDUID" != "not_stored" ]] && [[ "$STEAMCMDPSW" != "not_stored" ]]; then
+			script_update
+			script_update_mods
+		fi
 		script_update_github
 	fi
 }
@@ -1005,8 +1042,10 @@ script_timer_two() {
 		echo "$(date +"%Y-%m-%d %H:%M:%S") [$VERSION] [$NAME] [INFO] (Status) Server running." | tee -a "$LOG_SCRIPT"
 		script_enabled
 		script_logs
-		script_update
-		script_update_mods
+		if [[ "$STEAMCMDUID" != "not_stored" ]] && [[ "$STEAMCMDPSW" != "not_stored" ]]; then
+			script_update
+			script_update_mods
+		fi
 		script_update_github
 	fi
 }
@@ -1087,6 +1126,7 @@ script_install() {
 	echo ""
 	sudo useradd -m -g users -s /bin/bash $USER
 	echo -en "$USER_PASS\n$USER_PASS\n" | sudo passwd $USER
+	
 	echo ""
 	echo "You will now have to enter your Steam credentials. Exepct a prompt for a Steam guard code if you have it enabled."
 	echo ""
@@ -1103,7 +1143,17 @@ script_install() {
 			echo "Please try again."
 		fi
 	done
+	
 	echo ""
+	read -p "Do you want the script to store your Steam credentials in the script's config file for automatic updates? (y/n)" STEAM_STORE_CREDENTIALS_SETUP
+	STEAM_STORE_CREDENTIALS_SETUP=${STEAM_STORE_CREDENTIALS_SETUP:=n}
+	if [[ "$STEAM_STORE_CREDENTIALS_SETUP" =~ ^([nN][oO]|[nN])$ ]]; then
+		echo "Your Steam credentials WILL NOT be stored on this system. You will have to update your game manually when an update is released."
+		STEAM_STORE_CREDENTIALS="0"
+	else
+		echo "Your Steam credentials WILL be stored on this system. Updates will be installed automaticly when an update is released."
+		STEAM_STORE_CREDENTIALS="1"
+	fi
 	
 	echo ""
 	read -p "Enable beta branch? Used for experimental and legacy versions. (y/n): " SET_BETA_BRANCH_STATE
@@ -1213,7 +1263,11 @@ script_install() {
 	fi
 	
 	echo ""
+	echo "WARNING: script updates from github may include malicious code to steal any info the script uses to work, like Steam accound and password."
+	echo "Not saying i'm that kind of person that would do that but:"
+	echo "IF YOU DON'T TRUST ME, LEAVE THIS OFF FOR SECURITY REASONS!"
 	read -p "Enable automatic updates for the script from github? (y/n): " SCRIPT_UPDATE_CONFIG
+	SCRIPT_UPDATE_CONFIG=${SCRIPT_UPDATE_CONFIG:=n}
 	if [[ "$SCRIPT_UPDATE_CONFIG" =~ ^([yY][eE][sS]|[yY])$ ]]; then
 		SCRIPT_UPDATE_ENABLED="1"
 	else
@@ -1384,8 +1438,13 @@ script_install() {
 	script_install_tmux_config
 	
 	touch $SCRIPT_DIR/$SERVICE_NAME-config.conf
-	echo 'username='"$STEAMCMDUID" > $SCRIPT_DIR/$SERVICE_NAME-config.conf
-	echo 'password='"$STEAMCMDPSW" >> $SCRIPT_DIR/$SERVICE_NAME-config.conf
+	if [[ "$STEAM_STORE_CREDENTIALS" == "1" ]]; then
+		echo 'username='"$STEAMCMDUID" > $SCRIPT_DIR/$SERVICE_NAME-config.conf
+		echo 'password='"$STEAMCMDPSW" >> $SCRIPT_DIR/$SERVICE_NAME-config.conf
+	else
+		echo 'username=not_stored' > $SCRIPT_DIR/$SERVICE_NAME-config.conf
+		echo 'password=not_stored' >> $SCRIPT_DIR/$SERVICE_NAME-config.conf
+	fi
 	echo 'beta_branch_enabled='"$BETA_BRANCH_ENABLED" >> $SCRIPT_DIR/$SERVICE_NAME-config.conf
 	echo 'beta_branch_name='"$BETA_BRANCH_NAME" >> $SCRIPT_DIR/$SERVICE_NAME-config.conf
 	echo 'email_sender='"$POSTFIX_SENDER" >> $SCRIPT_DIR/$SERVICE_NAME-config.conf
@@ -1633,9 +1692,11 @@ EOF
 }
 
 #Do not allow for another instance of this script to run to prevent data loss
-if [[ $(pidof -o %PPID -x $0) -gt "0" ]]; then
-	echo "$(date +"%Y-%m-%d %H:%M:%S") [$NAME] [INFO] Another instance of this script is already running. Exiting to prevent data loss."
-	exit 0
+if [ "$1" != "-send_notification_start_initialized" ] || [ "$1" != "-send_notification_start_complete" ] || [ "$1" != "-send_notification_stop_initialized" ] || [ "$1" != "-send_notification_stop_complete" ] || [ "$1" != "-send_notification_crash" ]; then
+	if [[ $(pidof -o %PPID -x $0) -gt "0" ]]; then
+		echo "$(date +"%Y-%m-%d %H:%M:%S") [$NAME] [INFO] Another instance of this script is already running. Exiting to prevent data loss."
+		exit 0
+	fi
 fi
 
 if [ "$EUID" -ne "0" ] && [ -f "$SCRIPT_DIR/$SERVICE_NAME-config.conf" ]; then #Check if script executed as root, if not generate missing config fields
@@ -1660,28 +1721,26 @@ case "$1" in
 		echo -e "${CYAN}$NAME server script by 7thCore${NC}"
 		echo "Version: $VERSION"
 		echo ""
-		echo -e "${GREEN}start ${RED}- ${GREEN}Start the server${NC}"
-		echo -e "${GREEN}stop ${RED}- ${GREEN}Stop the server${NC}"
-		echo -e "${GREEN}restart ${RED}- ${GREEN}Restart the server${NC}"
-		echo -e "${GREEN}backup ${RED}- ${GREEN}Backup files, if server running or not.${NC}"
-		echo -e "${GREEN}autobackup ${RED}- ${GREEN}Automaticly backup files when server running${NC}"
-		echo -e "${GREEN}deloldbackup ${RED}- ${GREEN}Delete old backups${NC}"
-		echo -e "${GREEN}delete_save ${RED}- ${GREEN}Delete the server's save game with the option for deleting/keeping the server.json and SSK.txt files.${NC}"
-		echo -e "${GREEN}deloldsavefiles ${RED}- ${GREEN}Delete the server's save game (leaves the latest number files specefied in the script conf file.${NC}"
-		echo -e "${GREEN}change_branch ${RED}- ${GREEN}Changes the game branch in use by the server (public,experimental,legacy and so on).${NC}"
-		echo -e "${GREEN}rebuild_tmux_config ${RED}- ${GREEN}Reinstalls the tmux configuration file from the script. Usefull if any tmux configuration updates occoured.${NC}"
-		echo -e "${GREEN}rebuild_services ${RED}- ${GREEN}Reinstalls the systemd services from the script. Usefull if any service updates occoured.${NC}"
-		echo -e "${GREEN}rebuild_prefix ${RED}- ${GREEN}Reinstalls the wine prefix. Usefull if any wine prefix updates occoured.${NC}"
-		echo -e "${GREEN}disable_services ${RED}- ${GREEN}Disables all services. The server and the script will not start up on boot anymore.${NC}"
-		echo -e "${GREEN}enable_services ${RED}- ${GREEN}Enables all services dependant on the configuration file of the script.${NC}"
-		echo -e "${GREEN}reload_services ${RED}- ${GREEN}Reloads all services, dependant on the configuration file.${NC}"
-		echo -e "${GREEN}update ${RED}- ${GREEN}Update the server, if the server is running it wil save it, shut it down, update it and restart it.${NC}"
-		echo -e "${GREEN}update_mods ${RED}- ${GREEN}Update the server mods, if the server is running it wil save it, shut it down, update it and restart it.${NC}"
-		echo -e "${GREEN}update_script ${RED}- ${GREEN}Check github for script updates and update if newer version available.${NC}"
-		echo -e "${GREEN}update_script_force ${RED}- ${GREEN}Get latest script from github and install it no matter what the version.${NC}"
-		echo -e "${GREEN}status ${RED}- ${GREEN}Display status of server${NC}"
-		echo -e "${GREEN}install ${RED}- ${GREEN}Installs all the needed files for the script to run, the wine prefix and the game.${NC}"
-		echo -e "${GREEN}install_packages ${RED}- ${GREEN}Installs all the needed packages (Supports only Arch linux & Ubuntu 19.10 and onward)"
+		echo -e "${GREEN}-start ${RED}- ${GREEN}Start the server${NC}"
+		echo -e "${GREEN}-stop ${RED}- ${GREEN}Stop the server${NC}"
+		echo -e "${GREEN}-restart ${RED}- ${GREEN}Restart the server${NC}"
+		echo -e "${GREEN}-backup ${RED}- ${GREEN}Backup files, if server running or not${NC}"
+		echo -e "${GREEN}-autobackup ${RED}- ${GREEN}Automaticly backup files when server running${NC}"
+		echo -e "${GREEN}-deloldbackup ${RED}- ${GREEN}Delete old backups${NC}"
+		echo -e "${GREEN}-delete_save ${RED}- ${GREEN}Delete the server's save game with the option for deleting/keeping the server.cfg file${NC}"
+		echo -e "${GREEN}-change_branch ${RED}- ${GREEN}Changes the game branch in use by the server (public,experimental,legacy and so on.${NC}"
+		echo -e "${GREEN}-rebuild_tmux_config ${RED}- ${GREEN}Reinstalls the tmux configuration file from the script. Usefull if any tmux configuration updates occoured${NC}"
+		echo -e "${GREEN}-rebuild_services ${RED}- ${GREEN}Reinstalls the systemd services from the script. Usefull if any service updates occoured${NC}"
+		echo -e "${GREEN}-disable_services ${RED}- ${GREEN}Disables all services. The server and the script will not start up on boot anymore${NC}"
+		echo -e "${GREEN}-enable_services ${RED}- ${GREEN}Enables all services dependant on the configuration file of the script${NC}"
+		echo -e "${GREEN}-reload_services ${RED}- ${GREEN}Reloads all services, dependant on the configuration file${NC}"
+		echo -e "${GREEN}-update ${RED}- ${GREEN}Update the server, if the server is running it wil save it, shut it down, update it and restart it${NC}"
+		echo -e "${GREEN}-update_mods ${RED}- ${GREEN}Update the server mods, if the server is running it wil save it, shut it down, update it and restart it${NC}"
+		echo -e "${GREEN}-update_script ${RED}- ${GREEN}Check github for script updates and update if newer version available${NC}"
+		echo -e "${GREEN}-update_script_force ${RED}- ${GREEN}Get latest script from github and install it no matter what the version${NC}"
+		echo -e "${GREEN}-status ${RED}- ${GREEN}Display status of server${NC}"
+		echo -e "${GREEN}-install ${RED}- ${GREEN}Installs all the needed files for the script to run, the wine prefix and the game${NC}"
+		echo -e "${GREEN}-install_packages ${RED}- ${GREEN}Installs all the needed packages (Supports only Arch linux & Ubuntu 19.10 and onward)${NC}"
 		echo ""
 		echo -e "${LIGHTRED}If this is your first time running the script:${NC}"
 		echo -e "${LIGHTRED}Use the -install argument (run only this command as root) and follow the instructions${NC}"